Fix issues on travis-ci.org
Travis-CI was getting a coverage error.  pytest-cov only wants to cover
an entire directory, you can't ive it a single file.  I documented this
in Makefile near where COVER_ARG  is defined.

The Makefile itself is the modern version, with some fixes which will be
backported.

test/context.py adds project root directory to the path.